+++ CHANGELOG.md
@@ -207,4 +207,5 @@ Please add a _short_ line describing the PR you make, if the PR implements a spe
 
 ## Sprint (2023-02-17 - 2023-02-03)
 
+- Move Rotating log file maximum size and backup count to config variables ([#1388](https://github.com/ScilifelabDataCentre/dds_web/pull/1388))
 - Dependency: Werkzeug bumped to 2.2.3 ([#1387](https://github.com/ScilifelabDataCentre/dds_web/pull/1387))